Change mapping strategy to a pre-process workflow
This code walks back a handful of decisions about how to process
cron entries to now pre-process them into multiple possible entries
when a comma-separated list of options is encountered. This should
simplify date-generation later-on. It also means the conversion from
cron->struct->cron will currently be lossy if it was ever desired.
It could be considered a future optimization to bring that back if
ever converting back to cron.